docs: fix docstring typos and type errors in core modules
- PurpleAirAPIHelpers: fix 'NULL equivalent. i.e' → 'null equivalent (e.g., ...)' in sanitize_sensor_data_from_paa
- PurpleAirAPIHelpers: fix 'i.e' → 'i.e.,' in send_url_get_request separator param
- PurpleAirReadAPI: correct nwlng/nwlat/selng/selat parameter types from int to float (geographic coordinates require floating-point precision)
- PurpleAirMatterConverter: clarify pm25_to_aqi return type (returns float at boundary clamps, int otherwise)

+++ purpleair_api/PurpleAirReadAPI.py
@@ -130,15 +130,15 @@ class PurpleAirReadAPI:
                                        number of seconds. Using a value of 0 will match sensors of any age.
                                        Default value: 604800
 
-        :param (optional) int nwlng: A north west longitude for the bounding box. Use a bounding box to limit the sensors
+        :param (optional) float nwlng: A north west longitude for the bounding box. Use a bounding box to limit the sensors
                                      returned to a specific geographic area. The bounding box is defined by two points, a
                                      north west latitude/longitude and a south east latitude/longitude.
 
-        :param (optional) int nwlat: A north west latitude for the bounding box.
+        :param (optional) float nwlat: A north west latitude for the bounding box.
 
-        :param (optional) int selng: A south east longitude for the bounding box.
+        :param (optional) float selng: A south east longitude for the bounding box.
 
-        :param (optional) int selat: A south east latitude for the bounding box.
+        :param (optional) float selat: A south east latitude for the bounding box.
 
         :return dict | None: A python dictionary containing the payload response
         """
